Make Predictions on an elo.glm Object

Make Predictions on an elo.glm Object

# S3 method for elo.glm
predict(object, newdata, type = "response", ...)

# S3 method for elo.running
predict(object, newdata, running = TRUE, ...)

Arguments

object
An object of class "elo.glm".

newdata
A new dataset containing the same variables as the call that made object. If missing, the predicted win probabilities from object will be returned.

type
See predict.glm

...
Other arguments.

running
logical, denoting whether to use the running predicted values. Only makes sense if newdata is missing.

Value

A vector of win probabilities.

Details

Note that the "elo.glm.running" objects will use a model fit on all the data to predict.

Examples

data(tournament)
t1 <- head(tournament, -3)
t2 <- tail(tournament, 3)
results <- elo.glm(score(points.Home, points.Visitor) ~ team.Home + team.Visitor, data = t1,
  subset = points.Home != points.Visitor)
predict(results)

#>         1         2         3         4         5         6         7         8 
#> 0.2531029 0.1786407 0.7938802 0.7637434 0.6614671 0.5563559 0.6928515 0.6543741 
#>         9        10        11        12        13        14        15        16 
#> 0.1112190 0.9514122 0.6822634 0.6654358 0.5384049 0.9298593 0.5738479 0.9524142 
#>        17        18        19        20        21        22        23        24 
#> 0.5694956 0.9211923 0.9545990 0.8808230 0.6696690 0.9729000 0.4244605 0.8859010 
#>        25        26        27        28        29        30        31        32 
#> 0.6968958 0.9693763 0.9714558 0.9814906 0.7496476 0.7810671 0.8551242 0.9019272 
#>        33        34        35        36        37        38        39        40 
#> 0.8589852 0.3706649 0.9892661 0.8529083 0.8430390 0.9081528 0.4652261 0.3655728 
#>        41        42        43        44        45        46        47        48 
#> 0.4966376 0.8971005 0.8999737 0.8504990 0.2431403 0.9370506 0.6094407 0.2670447 

predict(results, newdata = t2)

#>         1         2         3 
#> 0.8337789 0.5976471 0.9398963