Make Predictions on an elo.markovchain Object

Make Predictions on an elo.markovchain Object

# S3 method for elo.markovchain
predict(object, newdata, ...)

Arguments

object
An object of class "elo.markovchain".

newdata
A new dataset containing the same variables as the call that made object. If missing, the predicted win probabilities from object will be returned.

...
Other arguments.

Value

A vector of win probabilities.

See also

Examples

data(tournament)
t1 <- head(tournament, -3)
t2 <- tail(tournament, 3)
results <- elo.markovchain(score(points.Home, points.Visitor) ~ team.Home + team.Visitor, data = t1,
  subset = points.Home != points.Visitor, k = 0.7)
predict(results)

#>         1         2         3         4         5         6         7         8 
#> 0.2658740 0.4410746 0.8602036 0.8505769 0.5085418 0.6927541 0.5358962 0.5164899 
#>         9        10        11        12        13        14        15        16 
#> 0.1983203 0.9041337 0.6656934 0.3824988 0.5070815 0.9458829 0.5344427 0.8922492 
#>        17        18        19        20        21        22        23        24 
#> 0.6572618 0.9036261 0.8276390 0.8497647 0.7664461 0.9623895 0.3829382 0.7663506 
#>        25        26        27        28        29        30        31        32 
#> 0.8594995 0.9320965 0.9678308 0.9324653 0.6390877 0.6568427 0.9132689 0.8285467 
#>        33        34        35        36        37        38        39        40 
#> 0.9107157 0.5360288 0.9778000 0.9462083 0.8454848 0.9137306 0.3840060 0.5681884 
#>        41        42        43        44        45        46        47        48 
#> 0.5374816 0.8503404 0.8462465 0.7685313 0.2986474 0.9680287 0.6582784 0.4425154 

predict(results, newdata = t2)

#>         1         2         3 
#> 0.6404343 0.7686262 0.9461136