Make Predictions on an elo.run Object

Make Predictions on an elo.run Object

# S3 method for elo.run
predict(object, newdata, ...)

# S3 method for elo.run.regressed
predict(object, newdata, regressed = FALSE, ...)

Arguments

object
An object of class "elo.run".

newdata
A new dataset containing the same variables as the call that made object. If missing, the predicted win probabilities from object will be returned.

...
Other arguments to be passed to elo.prob.

regressed
See the note on final.elos.

Value

A vector of win probabilities.

Examples

data(tournament)
t1 <- head(tournament, -3)
t2 <- tail(tournament, 3)
results <- elo.run(score(points.Home, points.Visitor) ~ team.Home + team.Visitor,
                   data = t1, k = 20)
predict(results)

#>  [1] 0.5000000 0.5000000 0.5000000 0.5000000 0.4712494 0.5287506 0.5287506
#>  [8] 0.4712494 0.5000000 0.5000000 0.5000000 0.5000000 0.5016550 0.4983450
#> [15] 0.4983450 0.5016550 0.5000000 0.5000000 0.5000000 0.5000000 0.5287506
#> [22] 0.5714631 0.4569335 0.5287506 0.5560040 0.5671044 0.5132404 0.5307154
#> [29] 0.5684153 0.5420635 0.4743640 0.4993501 0.5009797 0.4758315 0.4622916
#> [36] 0.5432464 0.4370737 0.5976816 0.5006344 0.4758579 0.5219654 0.3722182
#> [43] 0.4871266 0.4069742 0.4370169 0.5182687 0.3844961 0.4879447 0.5095624
#> [50] 0.3897233 0.5470091 0.4200704 0.3970499

predict(results, newdata = t2)

#> [1] 0.5076848 0.4379392 0.5874606